# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaraApi",
        "description": "LaraApi project API documents",
        "contact": {
            "email": "russell57260620@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/name": {
            "get": {
                "tags": [
                    "Name"
                ],
                "summary": "index",
                "description": "Display a listing of the name table.",
                "operationId": "getNameIndex",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Name"
                ],
                "summary": "store",
                "description": "Store a newly created name in name table.",
                "operationId": "postNameStore",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "head": {
                "tags": [
                    "Name"
                ],
                "summary": "index",
                "description": "Display a listing of the name table.",
                "operationId": "headNameIndex",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            }
        },
        "/api/name/{id}": {
            "get": {
                "tags": [
                    "Name"
                ],
                "summary": "show",
                "description": "Display the specified name with name id.",
                "operationId": "getNameShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Name ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Name"
                ],
                "summary": "update",
                "description": "Update the specified name in name table.",
                "operationId": "putNameUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Name ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Name"
                ],
                "summary": "destroy",
                "description": "Remove the specified name from name table.",
                "operationId": "deleteNameDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Name ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "head": {
                "tags": [
                    "Name"
                ],
                "summary": "show",
                "description": "Display the specified name with name id.",
                "operationId": "headNameShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Name ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Name"
                ],
                "summary": "update",
                "description": "Update the specified name in name table.",
                "operationId": "patchNameUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Name ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    },
...
}